+++ spec/filters/unsupported/freeze.rb
@@ -45,4 +45,6 @@ opal_filter "freezing" do
   fails "String#clone copies frozen state"
   fails "String#initialize with no arguments does not raise an exception when frozen"
   fails "StringScanner#initialize returns an instance of StringScanner"
+  fails "A singleton method definition raises RuntimeError if frozen"
+  fails "A method definition inside a metaclass scope raises RuntimeError if frozen"
 end